fast/regex/regexp-cross-frame-callable.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

function doTest() {
  re = frames[0].getRegExp();
  shouldBe("re('a')", "['a']");
}

var successfullyParsed = true;
</script>
<body onLoad="doTest()">
<p id="description"></p>
<div id="console"></div>
<iframe src="resources/regexp-cross-frame-callable-child.html"></iframe>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
